patch: Rename chat command reping to links
