--- src/meetbot/features/chat-command-reping.ts
@@ -1,22 +1,0 @@
-import { Bot } from '..';
-import { postToChatJob } from '../google-meet-helpers';
-
-/**
- * Resend important messages
- */
-export const attach = (bot: Bot) => {
-	let chatTranscriptUrl: string | null;
-
-	bot.on('chat_transcript_doc_ready', (data) => {
-		chatTranscriptUrl = data.transcriptUrl;
-	});
-
-	bot.on('chat', ({ text }) => {
-		if (text === '/reping') {
-			console.log('Reping requested');
-			bot.addJob(
-				postToChatJob(`Resending chat transcript: ${chatTranscriptUrl}`),
-			);
-		}
-	});
-};
